requirements.txt
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
idna==2.8
python-box==4.0.4
requests==2.22.0
ruamel.yaml==0.16.7
ruamel.yaml.clib==0.2.0
toml==0.10.0
urllib3==1.25.8
